refactor(was): package 분리
