--- src/main/java/controller/HomeControllerApis.java
@@ -1,19 +1,0 @@
-package controller;
-
-import org.springframework.http.HttpMethod;
-import utils.Api;
-
-public enum HomeControllerApis {
-    HELLO_WORLD_API(new Api("/", HttpMethod.GET)), INDEX_API(new Api("/index.html", HttpMethod.GET));
-
-    private final Api api;
-
-    HomeControllerApis(Api api) {
-        this.api = api;
-    }
-
-    public Api getApi() {
-        return api;
-    }
-
-}
