--- src/main/java/controller/MyController.java
@@ -1,11 +1,0 @@
-package controller;
-
-import java.io.DataOutputStream;
-import utils.request.MyRequest;
-
-public interface MyController {
-
-    boolean canHandle(MyRequest myRequest);
-
-    void handle(MyRequest myRequest, DataOutputStream dataOutputStream);
-}
