+++ src/main/java/controller/infra/MyController.java
@@ -1,0 +1,11 @@
+package controller.infra;
+
+import java.io.DataOutputStream;
+import utils.request.MyRequest;
+
+public interface MyController {
+
+    boolean canHandle(MyRequest myRequest);
+
+    void handle(MyRequest myRequest, DataOutputStream dataOutputStream);
+}
