--- src/main/java/dao/UserDao.java
@@ -1,7 +1,0 @@
-package dao;
-
-import entity.User;
-
-public interface UserDao {
-    User save(User user);
-}
