--- src/main/java/dao/UserDaoImpl.java
@@ -1,18 +1,0 @@
-package dao;
-
-import entity.User;
-import repository.UserRepository;
-
-public class UserDaoImpl implements UserDao {
-
-    private final UserRepository userRepository;
-
-    public UserDaoImpl(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public User save(User user) {
-        return userRepository.save(user);
-    }
-}
