--- src/main/java/db/UserDataBase.java
@@ -1,29 +1,0 @@
-package db;
-
-import entity.User;
-import java.util.Collection;
-import java.util.HashMap;
-import java.util.Map;
-
-public class UserDataBase {
-
-    private static final Map<String, User> users = new HashMap<>();
-
-    private UserDataBase(){}
-
-    public static User insert(User user){
-        users.put(user.getUserId(), user);
-        return users.get(user.getUserId());
-    }
-    public static User findByUserId(String userId){
-        return users.get(userId);
-    }
-    public static Collection<User> findAll(){
-        return users.values();
-    }
-
-    public static void deleteAll(){
-        users.clear();
-    }
-
-}
