--- src/main/java/repository/UserRepository.java
@@ -1,7 +1,0 @@
-package repository;
-
-import entity.User;
-
-public interface UserRepository {
-    User save(User user);
-}
