--- src/main/java/repository/UserRepositoryImpl.java
@@ -1,12 +1,0 @@
-package repository;
-
-import db.DataBase;
-import entity.User;
-
-public class UserRepositoryImpl implements UserRepository {
-    @Override
-    public User save(User user) {
-        DataBase.addUser(user);
-        return user;
-    }
-}
