--- src/main/java/service/UserService.java
@@ -1,9 +1,0 @@
-package service;
-
-import dto.UserCreateDto;
-import entity.User;
-
-public interface UserService {
-
-    User create(UserCreateDto userCreateDto);
-}
