--- src/main/java/service/UserServiceImpl.java
@@ -1,20 +1,0 @@
-package service;
-
-import dao.UserDao;
-import dto.UserCreateDto;
-import entity.User;
-
-public class UserServiceImpl implements UserService {
-
-    private final UserDao userDao;
-
-    public UserServiceImpl(UserDao userDao) {
-        this.userDao = userDao;
-
-    }
-
-    @Override
-    public User create(UserCreateDto userCreateDto) {
-        return userDao.save(userCreateDto.toEntity());
-    }
-}
