+++ src/main/java/user/dao/UserDao.java
@@ -1,0 +1,7 @@
+package user.dao;
+
+import user.User;
+
+public interface UserDao {
+    User save(User user);
+}
