+++ src/main/java/user/dao/UserDaoImpl.java
@@ -1,0 +1,18 @@
+package user.dao;
+
+import user.User;
+import user.repository.UserRepository;
+
+public class UserDaoImpl implements UserDao {
+
+    private final UserRepository userRepository;
+
+    public UserDaoImpl(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    public User save(User user) {
+        return userRepository.save(user);
+    }
+}
