+++ src/main/java/user/repository/UserRepository.java
@@ -1,0 +1,7 @@
+package user.repository;
+
+import user.User;
+
+public interface UserRepository {
+    User save(User user);
+}
