+++ src/main/java/user/service/UserService.java
@@ -1,0 +1,9 @@
+package user.service;
+
+import user.User;
+import user.dto.UserCreateDto;
+
+public interface UserService {
+
+    User create(UserCreateDto userCreateDto);
+}
