+++ src/main/java/utils/request/MyHeaders.java
@@ -7,7 +7,7 @@ public class MyHeaders extends MyRequestMap{
     @Override
     public String get(String key) {
         if(key.equals(HttpHeaders.ACCEPT)){
-            String value = getOrDefault(key, MediaType.TEXT_HTML_VALUE);
+            String value = super.getOrDefault(key, MediaType.TEXT_HTML_VALUE);
             return value.split(",")[0];
         }
         return super.get(key);
