+++ src/test/java/db/UserDataBaseTest.java
@@ -2,7 +2,8 @@ package db;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import entity.User;
+import user.User;
+import infra.db.UserDataBase;
 import java.util.Collection;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
